Adaptando a Sirhus

+++ src/Controller/Quinque/CategoriaController.php
@@ -11,31 +11,29 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
-#[Route('/categoria')]
+
+#[Route('/categoria', name: 'intranet_quinque_admin_categoria_')]
 class CategoriaController extends AbstractController
 {
-    private EntityManagerInterface $entityManager;
-    private CategoriaRepository $categoriaRepository;
-
     public function __construct(
-        EntityManagerInterface $entityManager,
-        CategoriaRepository $categoriaRepository,
+        private readonly MessageGenerator $generator,
+        private readonly CategoriaRepository $categoriaRepository,
     ) {
-        $this->entityManager = $entityManager;
-        $this->categoriaRepository = $categoriaRepository;
     }
 
-    #[Route('/', name: 'quinque_categoria_index', methods: ['GET'])]
+    #[Route('/', name: 'index', methods: ['GET'])]
     public function index(): Response
     {
+        $this->denyAccessUnlessGranted('admin');
         return $this->render('intranet/quinque/admin/categoria/index.html.twig', [
             'categorias' => $this->categoriaRepository->findAll(),
         ]);
     }
 
-    #[Route('/new', name: 'quinque_categoria_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
-    {
+    {        
+        $this->denyAccessUnlessGranted('admin');
         $categoria = new Categoria();
         $form = $this->createForm(CategoriaType::class, $categoria);
         $form->handleRequest($request);
@@ -53,17 +51,19 @@ class CategoriaController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'quinque_categoria_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'show', methods: ['GET'])]
     public function show(Categoria $categoria): Response
     {
+        $this->denyAccessUnlessGranted('admin');
         return $this->render('intranet/quinque/admin/categoria/show.html.twig', [
             'categoria' => $categoria,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'quinque_categoria_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Categoria $categoria): Response
     {
+        $this->denyAccessUnlessGranted('admin');
         $form = $this->createForm(CategoriaType::class, $categoria);
         $form->handleRequest($request);
 
@@ -79,9 +79,10 @@ class CategoriaController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'quinque_categoria_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'delete', methods: ['POST'])]
     public function delete(Request $request, Categoria $categoria): Response
     {
+        $this->denyAccessUnlessGranted('admin');
         if ($this->isCsrfTokenValid('delete'.$categoria->getId(), $request->request->get('_token'))) {
             $this->entityManager->remove($categoria);
             $this->entityManager->flush();
